Remove variable-pitch-mode from org-mode.
Unfortunately, I just can't get it to work nicely with org-indent and
visual-fill-column-mode.

With org-indent, the indentation is not enough because a proportional space is
much less than other characters.  The indent ends up not being enough to line up
with the headlines.  I tried customizing the `org-indent` face to use a fixed
width font, which works, for headlines, but plain lists still get wonky
indentation.

With visual-fill-column-mode and `org-indent` customized to a fixed width font,
it ends up indenting too much.  Apparently this happens with visual-line-mode,
too, and there's a fix (from 2009; is my org really that old?).

So the combination of org-indent, customizing org-indent to use a fixed width
font, and visual-fill-column-mode makes variable-pitch-mode work (assuming the
fix works, but I haven't tried it), but there's still the problem that plain
lists don't indent correctly.  Maybe if I could figure that out I could go back
to using variable-pitch-mode globally for org-mode.

For now, I only have variable-pitch-mode enabled in my zettelkasten, which is
where I'm writing prose and really want the proportional width font anyway.  I
also don't think I'll be using a bunch of nested plain lists, so I think that's
an appropriate compromise.