Upgrade magit
My issues with password prompts were because I upgraded openssh to 8.0 which
changed the password prompt slightly, and magit was not detecting the prompt.
This was fixed a year ago in magit 0.91.0, but that has not been release to
melpa-stable, so I had to add melpa as a repository, set a default pin to
melpa-stable, and upgrade magit.