Fix exwm prefix keys on login
On login, exwm prefix keys would not work for X windows until I reset using
pjs-reset.  I narrowed down the part of pjs-reset that was fixing the prefix
keys to the loading of my init file, and from there to the use-package form for
exwm and the customization for exwm-input-prefix-keys.

It turns out that exwm-input-prefix-keys has a custom setter. The first time my
init file was loaded exwm-input-prefix-keys was customized before the setter was
defined, and it was setting an invalid value.  The setter was unwrapping array
values, my value had arrays, and the arrays were not getting unwrapped.

When I ran pjs-reset and it loaded my init file a second time, the custom setter
was defined, and it was correctly unwrapping array values.

In the end, I was attempting to set a value that I copied from my custom.el,
which was just the default value, so there's no reason to even set
exwm-input-prefix-keys. ¯\_(ツ)_/¯