Define pjs-org-insert-created-property-p before loading pjs-org
pjs-org is deferred, and a file local or dir local will try to assign a value to
pjs-org-insert-created-property-p before pjs-org is loaded, and that triggers a
nasty warning and confirmation.

If we define it eagerly and give it a safe value predicate, then we can assign
the var even before pjs-org is loaded.